Page Submissions Imported and Correction of Minor Mistakes
The text of the page submissions has been imported. For the The wiki page "submissions," HTML elements and wiki tables are used. However, they could not be copied adequately due to limitations posed by markdown. Some sections of the markdown text need to be reworked because they are a terrible read for multiple reasons.