Use old strategy for datepicker import (npm, not CDN)
We use the npm-again, which is working again suddenly (?!)
The only thing that changed is basically that we don't explicitly require
the module anymore in application.js just to set its locale.
Also, renamed the import to a random string so that it's obvious
this is not a real module import, but rather an ugly fix. Also added
a small comment regarding this.
We also reimport the datepicker css in application.sass