Bootstrap5 (#463)
* added bs5
reworked media accordions and all jumbotrons usages

* custom-chechbox to form-checkbox

data-dismiss="modal" to data-bs-dismiss="modal"

fixed footer placement for devise views

* Renamed .ml-* and .mr-* to .ms-* and .me-*.
Renamed .pl-* and .pr-* to .ps-* and .pe-*.
Renamed .border-left and .border-right to .border-start and .border-end.
Renamed .text-left and .text-right to .text-start and .text-end.
Navbars now require a container within (to drastically simplify spacing requirements and CSS required).

* Continued working on this change:
Dropped form-specific layout classes for our grid system.

* fix home layout to be consistent with prior version

* "custom-control-*" to "form-check-*"
"form-checkbox" to "form-check"
"badge badge-light" to "badge badge-light text-dark"
started adding "form-label" to form-labels and removing "form-group"

* add form-label class to most form-label
"form-group" -> "mb-3"

* add form-label class to most form-label v2
fixed helpdesks, added popover js to "bootstrap_modal_turbolinks_fix.coffee"

* fix media-search form

* fixed bootstrap issues
modals
some dropdowns
navs
collapsibles

* fixed modals
accordion
helpdesks

* change js_compressor uglifier to terser

* fix issues

* changed from-control to form-select
for a lot of select fields
fixed a few helpdesks

* fixing href underlines
fix media search form
fix comment form-checks
fix bg-gray buttons

* fix new medium starting selected sort

* add data-masonry to fix card grid used for lecture content display

* Merge branch mampf-next into bootstrap5

* Fix missing "bg-" instead of "badge-"

* Fix link underline in badges

* Fix content tags color

* Remove padding from login remember_me checkbox

* Restore card header and card footer padding

* Remove left-over code

* Remove underline upon button hovering

* Remove text decoration from nav links

* Add background color to tags in admin view for lesson

* Replace lectures edit card columns by masonry

* Replace seminar card-column by masonry

* Replace structure (example) card-columns with masonry

* Remove left-over card-columns breakpoint class

* Restore "data-parent" instead of "data-bs-parent"

data-parent is a custom attribute and not related to bootstrap

* Revert "Restore "data-parent" instead of "data-bs-parent""

This reverts commit 74cdf0bfd73cdbef7a3e5dd7a28b64071cc66d35.

* Use data-parent for custom attribute

We use a custom attribute "data-parent", not to confuse with the new
"data-bs-parent" attribute of bootstrap (since v5.0).

* Revert tom-select bootstrap4 accordion fix

Fix originally introduced in: 7db4b45
Also see this issue: https://github.com/orchidjs/tom-select/issues/454

* Fix closing button of sponsors modal

(data-bs-dismiss instead of data-dismiss)

* Restyle sponsors modal close button

* Fix helpdesk method not found in commontator view

Should eventually be replaced by helper module inclusion after #445.

* Use cdn version of jquery-datetimepicker

* Remove autocomplete for datetimepicker input field

* Update chartjs-adapter-moment version

* Remove datetimepicker from application.js (webpacker)

* Remove unnecessary coffeescript hello world file

* Restore empty application.sass to make webpacker happy

* Remove left padding of preview toggle (commontator)

* Remove unused dependencies in yarn.lock

(Deletion was produced by webpacker automatically)

* Fix positioning of external link label

* Replace more custom-radio classes with form-check

custom-control custom-radio -> form-check
custom-control-inline -> form-check-inline

* Add padding to "comments" in publish medium modal

* Use .btn-close instead of .close for close buttons

* Get rid of "&times" in span for closing button

Bootstrap5 now uses a background image (embedded SVG) to display the
"x" on the closing buttons.

* Fix closing button on sponsors modal

* Replace custom-control-label by form-label

* Remove "row" class for language section

This makes sure the language options are shown in one line.

* Use form-check-label instead of form-label

We use this class for labels next to checkboxes to avoid too much
upper/lower spacing.

* Readd hello world coffee for webpacker to be happy

* Use old strategy for datepicker import (npm, not CDN)

We use the npm-again, which is working again suddenly (?!)
The only thing that changed is basically that we don't explicitly require
the module anymore in application.js just to set its locale.
Also, renamed the import to a random string so that it's obvious
this is not a real module import, but rather an ugly fix. Also added
a small comment regarding this.
We also reimport the datepicker css in application.sass

* Init datetimepicker on pageload

This allows us to directly toggle it, so that the user does not have
to click twice in the input field for the picker to show up.

* Get back to "css"/"myLib" import for datepicker

The random names apparently did not work either
(it worked first, again, due to docker caching).

* Get rid of unnecessary divs in quizzes view

* Wrap quizzes in full-width overflow-hidden container

This is to prevent the scrollbar showing up randomly when invoking the
jquery .show() method (via .coffee files).

* Fix missing "form-check-label" CSS class

* Remove unwanted "text-end" css class

* Remove padding around media closing buttons

* Fix coloring of tag badges

* Replace some form-control classes by form-select

This is to show the additional dropdown arrow in the respective field.

* Bunk replace tag css classes

to have consistent styling (e.g. light gray background,
on hover dark background with white font)

* Force no underline of dropdown menu items

* Replace remnant "custom-control" by "form-check"

(referring to the CSS class names)

* Remove left-over console.log()

* Delete simple_spec.cy.js

This file was already deleted in commit a0557b8 in mampf-next.
However, due to unknown reasons, during a merge of mampf-next
into this branch, I reintroduced the file [1]. This is why this commit
deletes it again.

[1] https://github.com/MaMpf-HD/mampf/commit/843854fe8317103b12b78e903a10ae17d59de77c#diff-f971a7862ba60ddb68800528aa4209773ddb3622a1507a123b0afc5aa3271a98

* Remove unwanted `testing_spec.cy.js`

Reason: see commit message in commit d0f6541

* Remove dockerfile remnant

This change was actually done in [1] but somehow got reintroduced in this branch again, so we remove it here manually.

[1] https://github.com/MaMpf-HD/mampf/commit/a0557b868b397d85d72fea4d5b73566c192393a7#diff-6d58d1f5718d589c6e138b97441c26ea8caaa3e0973f7c2c73ba7a64d600aec7

* Fix layout of long announcements

* Fix duplicate close button

* Fix "du" vs. "Du" with variants e.g. "dich" etc.

* Fix badge color not applied to all elements

* Fix badge color in admin search view

* Fix badge color in related tags info popup

* Fix tag color on light background

* Fix button underline in "related media"

* Fix uneven alignment of close medium button

* Add missing "show media" button to seminar content

* Add missing helpdesk text for lesson details

* Visually separate multiple announcements

* Make button colors on lecture cards consistent

text-secondary (gray) instead of text-dark (black)

* Fix seminar content card not taking full width

We replace its id by 'lecture-content-card' as this is the key
used in the respective lectures.coffee file to
add or remove the class 'col-xxxl-9'

* Add megaphone icon in front of every announcement

(on login page)

---------

Co-authored-by: Splines <dominic-plein@gmx.de>