Do not load masonry grid "asyncly" and fix initialization (#523)
"turoblinks:load" does not work well when scripts are loaded
asynchronously. Therefore, we do not load the masaonry grid
asynchronously anymore, see https://stackoverflow.com/a/50153681/9655481

Another problem (or the main problem) is that we've used plain HTML
to init Masonry options. However, the grid is populated dynamically
with Rails content. Therefore, we need to initialize it only when the
page is fully loaded. For this, we use JS and the respective CSS classes
on the elements. I called the class ".masonry-grid".