Redo the Rubocop config (#564)
* Shorten RuboCop config

Many cops were explicitly enabled. We don't necessarily need to do that
since the defaults of RuboCop are always included. So we just specify
where we deviate from RuboCops default values.

* Add comment regarding "NewCops"

* Add "NewCops: enable"

This has as consequence that whenever we upgrade RuboCop, we should run
checks on the whole codebase to see if new cops somehow don't work with
our code. We can then discuss if we maybe want to explicitly disable a new
cop.

* Update RuboCop related packages in Gemfile

* Add WordArray & EmptyMethod enforced styles

* Sort RuboCop keys alphabetically & improve file comment