Fix erroneous media search (#593)
* Warn about too long GitHub commit messages (#586)

* Fix comment status (#585)

* Reapply first fix for Reader/Media

See discussion on #574 for further details.
Previous PR for this was #576, closed in favor of this one
as this directly branches off the new "dev" branch.

* Correctly show latest post (might be current_user's comment)

* Fix update of unread comments logic in comments controller

* Fix update icon logic and latest post comment

* Simplify latest comment logic

* Improve code comments

* Further improve comments

* Fix wording in comment

* Fix construction of media array & use `.blank?` instead of `.empty?`

* Migrate `unread_comments` flag (fix inconsistencies) (#587)

* Add dummy migration

* Implement migration for unread comment flag

* Remove unnecessary comment

* Declare migration as not idempotent

* Use array.length instead of counting

* Throw error to prevent revert of migration

* Fix severe flaws in unread comments migration

* Simplify Reader retrieval

* Use the more explicit `.nil?` method

* Update migration date

* Fix annoying bug: don't use `.select!` but `.select`

* Polish migration

e.g. update comment, more suitable name for the method etc.

* Rename method according to #585

* Use `warn` log level for migration (#588)

* Rename get_statistics.coffee file to statistics.coffee (#591)

This is to reflect the corresponding renaming of the route
due to rubocop.

* fix behaviour of media search

* Disable OR/AND buttons if "all" tags is enabled

(inside the media search, right now only frontend change)

* Use ids autogenerated by Rails

This is such that we can still click on the label to select
the respective radio button, instead of only being able
to click on the radio button itself.

* Disable AND/OR initially (until user deselects "all tags")

* Don't restrict media search if "all" tags is enabled.

If the "all tags" option is enabled in the search, we should allow the
results to include *any* tag. This is automatically the case if we just
don't add any restriction regarding the tag ids in the SOLR search.

This change accompanies the frontend change to disable the AND/OR
buttons if the "all" button is selected meaning the user wants to search
for all tags. See #593 for more details.

---------

Co-authored-by: Splines <37160523+Splines@users.noreply.github.com>
Co-authored-by: Splines <dominic-plein@gmx.de>