Remove badges that count questionnaires
I removed them since these badges are usually used when there are a lot
of elements. The number of questionnaires is probably rather limited
and will unlikely surpass 10 in total. Therefore, I remove these badges
to ease quick visual parsing of the page with the eye without being
distracted by the very prominent and contrasty-color badges.