Remove webpacker and sass-rails & add basic vite_rails config
See this blog post by the creator of vite-rails:
https://maximomussini.com/posts/a-rubyist-guide-to-vite-js#getting-started

I've added the vite_rails gem, installed it, then ran:
bundle exec vite install

(Note that I removed the app/javascript folder beforehand, only locally,
as we don't need it.)

This is to get a basic setup.
Nothing is expected to work at this moment.