Remove core-js dependency
Just including it does nothing and there is no occurrence in our code for it.
Therefore, we can just as well get rid of it. As we don't even use some very
advanced, modern JS features, a polyfill is likely not really necessary.

Should we find some weird browser behavior, it's easy to get back to it.