Set locale also for Commontator (comments) (#938)
The `CommentsController` does not extend the usual
`ApplicationController`, but the `Commontator::ApplicationController`,
which didn't use our `set_locale` method. Therefore, I've extracted this
part to a new concern `LocaleSetter` that is included in both
application controllers. Note that the order matters, i.e. at which line
we write `include LocaleSetter`.

Due to this, the Playwright test that posted comments was flaky
beforehand because the user locale was not set deterministically. This
should work without any problems now.