Fix table of contents not loaded for videos (#943)
The table of content (TOC) on the right side of the videoplayer should
be automatically shown whenever there is at least one item in it.
However, this was not the case anymore: it stayed closed.


https://github.com/user-attachments/assets/6aa232df-9508-4386-a34f-ae94c2cc77f3

The problem was that the [`onloadedmetadata`
event](https://developer.mozilla.org/en-US/docs/Web/API/HTMLMediaElement/loadedmetadata_event)
was not fired since by the time we registered the handler, the video was
already loaded (thanks to quicker loading due to Vite).

Also added a Playwright test to avoid regressions.