Muesli: PR 2.1 - Schema and core models for registration (#922)
This is PR 2.1 from the [PR
roadmap](https://mampf-hd.github.io/mampf/features/implementation-prs.html)
for the MÜSLI integration. It contains the core Active Record models for
registrations and migrations to manifest them in the database.
The models created are:

- `Registration::Campaign`
- `Registration::Item`
- `Registration::UserRegistration`
- `Registration::Policy`

The PR contains factories for these models and unit tests that
demonstrate that these factories work as well. Note that all the changes
in this PR are additive.

For more details on the meaning of these models, I refer to the
[corresponding
chapter](http://localhost:3004/features/02-registration.html) of the
architecture book.

---------

Co-authored-by: Splines <37160523+Splines@users.noreply.github.com>