Fix manuscript upload content extraction (#949)
This supersedes #948.

A user reported that in production, the extraction of metadata from
scripts produced using the `mampf.sty` Latex package does work anymore.
This can be reproduced in development using the manuscript the user
provided (for lectures that use "using a manuscript generated by the
MaMpf LaTeX package" for content determination which can be set under
preferences).

<img width="1156" height="206" alt="Screenshot 2025-11-28 144843"
src="https://github.com/user-attachments/assets/163279a4-fe85-478c-a955-8dc24763bbfa"
/>


**The problem was an invalid statement in the monster `upload.coffee`
file. It is likely that this issue was already present before the latest
release.** (Note that only the manuscript upload was affected, other
uploads should still work)

Note that we should probably rethink the logic here a bit in the future.
What happens is that the user uploads the PDF, the PDF uploader class
extracts the metadata, then sends it back to the frontend, then the
frontend sends that Metadata back to the backend during the actual
"Save" action.