Disable turbo prefetching for questionnaire links (#959)
In the index view for vignettes questionnaires, a completion status is
shown for the questionnaires.

<img width="1020" height="493" alt="Screenshot 2025-12-04 135016"
src="https://github.com/user-attachments/assets/5c739d50-70d0-4cad-af22-ac336fe1ab74"
/>


With our new release, the setup uses more aggressive prefetching such
that whenever a user only hovered over a questionnaire link, the take
action was triggered in the background and in turn, the questionnaire
was saved as having been begun and the text in the badge then changed
from "Take" to "Complete". This PR disables prefetching for the
questionnaire links.

Co-authored-by: Splines <37160523+Splines@users.noreply.github.com>