Fix vignettes paths in LecturesController and navbar layout (#968)
A user reported that vignettes that had the `organizational` flag set to
`true` would trigger an exception when the show action was called. Trace
shows

```
[bd2c2bae-4aae-4369-b888-14cf39a37740] Started GET "/lectures/257" for 129.206.26.148 at 2025-12-11 11:13:35 +0000
11.12.2025 12:13:35
[bd2c2bae-4aae-4369-b888-14cf39a37740] Processing by LecturesController#show as HTML
11.12.2025 12:13:35
[bd2c2bae-4aae-4369-b888-14cf39a37740]   Parameters: {"id" => "257"}
11.12.2025 12:13:35
[bd2c2bae-4aae-4369-b888-14cf39a37740] Completed 500 Internal Server Error in 8ms (ActiveRecord: 1.2ms (3 queries, 0 cached) | GC: 0.0ms)
11.12.2025 12:13:35
[bd2c2bae-4aae-4369-b888-14cf39a37740]   
11.12.2025 12:13:35
[bd2c2bae-4aae-4369-b888-14cf39a37740] NoMethodError (undefined method 'organizational_path' for an instance of LecturesController):
11.12.2025 12:13:35
[bd2c2bae-4aae-4369-b888-14cf39a37740]   
11.12.2025 12:13:35
[bd2c2bae-4aae-4369-b888-14cf39a37740] app/controllers/lectures_controller.rb:23:in 'LecturesController#show'
```

This bug is unrelated to the new release, `organizational_path` was
never a valid path, it should be `lecture_organizational_path`
(apparently the case that this flag was `true` was never checked and
never occured in production). I fixed that. When I had fixed that, it
turned out that in turn another error was now triggered, this having to
do with changed paths for the vignettes assets due to the new folder
structure. I fixed that as well.

Co-authored-by: Splines <37160523+Splines@users.noreply.github.com>