Remove CoffeeScript for KaTex (#952)
This PR migrates the coffeescript file for KaTeX rendering to JS. It
also adds an Playwright end-to-end tests to ensure math formulas render
correctly in comments.

We still have Katex in the `_head.html.erb` at this moment as many views
rely on it. Getting rid of it is something for a future PR.

---------

Co-authored-by: Splines <37160523+Splines@users.noreply.github.com>